+++ CHANGELOG.md
@@ -2,6 +2,7 @@
   * Commit: 
   * New Commands
   * ParamsMap
+  * Get command params
 ## v0.3.7 - Application Packages 
   * Commit: 2026-04-07 13:50
   * Fail functions